+++ src/components/mod.rs
@@ -1,1 +1,3 @@
+mod typography;
 
+pub use self::typography::Typography;