+++ CHANGELOG.md
@@ -7,5 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-- **added**: Add Material Design system (including colors, shapes, motion and elevation).
+- **added**: Add Material Design system (including colors, shapes, motion, elevation and typography).
 - **added**: Add `func:use_material_theme` and `func:set_material_theme` for managing app theme.
+- **added**: Add `trait:LaunchConfigExt`, which provides `func:with_roboto` for `struct:LaunchConfig`, to add the `Roboto` font to the application.
+- **added**: Add `component:Typography`.
+- **added**: Add example of using `component:Typography`.